start adding SC stuff
