+++ mcp/include/star_camera_solutions.h
