+++ mcp/include/star_camera_transmit.h
