+++ mcp/sensors/CMakeLists.txt
@@ -7,6 +7,8 @@ add_library (mcp_sensors STATIC
     ${CMAKE_CURRENT_SOURCE_DIR}/inclinometer.c
     ${CMAKE_CURRENT_SOURCE_DIR}/magnetometer.c
     ${CMAKE_CURRENT_SOURCE_DIR}/star_camera_receive.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/star_camera_solutions.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/star_camera_transmit.c
     ${CMAKE_CURRENT_SOURCE_DIR}/xsc_network.c
     ${CMAKE_CURRENT_SOURCE_DIR}/xsc_protocol.c
 )
